playlist processing refactor
